Switched to time-managed deployment/undeployment
Before, getPosition was relied on to return the current position of the servos, which apparently it does not do.

There were also a few constants missing and a typo in a used constant in one of the state functions.

System.currentTimeMillis is assumed to return a float, although I haven't been able to test this. It also might not be imported.