Implemented PID movement using PIDVA follower
User input and previous pose are now converted into a target pose, velocity, and acceleration suitable for a goto function I added to SampleMecanumDrive, which calls a special public overload of the internalUpdate function in the HolonomicPIDVAFollower that returns a Pose2d that the goto function then forwards to the drive

Had to translate a few files from kotlin to get it to compile (Sorry @ustuej). It seems to run in my emulator now.